+++ src/amuse/ext/job_server.py
@@ -30,7 +30,7 @@ derived from the main module. Easy way to achieve this is to import them from a 
 """
 from amuse.rfi.core import *
 import cPickle as pickle
-from amuse.rfi.channel import AsyncRequestsPool
+from amuse.rfi.async_request import AsyncRequestsPool
 import inspect
 from collections import deque
 import threading
